start separating out listOps
